# archive/monolith-v1/openapi/ignyte-anchor.yaml
openapi: 3.0.3
info:
  title: Ignyte Anchor API
  version: 1.0.0
  description: Capability issuance and action verification API for Ignyte Anchor.
servers:
  - url: https://api.ignyteanchor.com
    description: Production
  - url: http://localhost:8080
    description: Local development
paths:
  /healthz:
    get:
      summary: Service health
      responses:
        '200':
          description: Service status
          content:
            application/json:
              schema:
                type: object
                required: [status, name, api]
                properties:
                  status:
                    type: string
                  name:
                    type: string
                  api:
                    type: string
  /v1/capabilities:
    post:
      summary: Issue a signed capability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueCapabilityRequest'
      responses:
        '201':
          description: Signed capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueCapabilityResponse'
        '400':
          description: Validation error
  /v1/actions/verify:
    post:
      summary: Verify an action envelope against a capability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyActionRequest'
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationResult'
        '400':
          description: Validation error
components:
  schemas:
    CapabilityConstraints:
      type: object
      required:
        - resource_limits
        - spend_limits
        - api_scopes
        - rate_limits
        - environment_constraints
      properties:
        resource_limits:
          type: object
          additionalProperties:
            type: integer
            format: int64
        spend_limits:
          type: object
          additionalProperties:
            type: integer
            format: int64
        api_scopes:
          type: array
          items:
            type: string
        rate_limits:
          type: object
          additionalProperties:
            type: integer
            format: int64
        environment_constraints:
          type: array
          items:
            type: string
    Capability:
      type: object
      required:
        - version
        - capability_id
        - issuer_id
        - agent_id
        - allowed_actions
        - constraints
        - issued_at
        - expires_at
        - nonce
        - signature
      properties:
        version:
          type: integer
        capability_id:
          type: string
        issuer_id:
          type: string
        agent_id:
          type: string
        allowed_actions:
          type: array
          items:
            type: string
        constraints:
          $ref: '#/components/schemas/CapabilityConstraints'
        issued_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        nonce:
          type: string
        signature:
          type: string
    ConstraintEvidence:
      type: object
      required:
        - resource_usage
        - spend_usage
        - rate_usage
        - environment
        - api_scope
      properties:
        resource_usage:
          type: object
          additionalProperties:
            type: integer
            format: int64
        spend_usage:
          type: object
          additionalProperties:
            type: integer
            format: int64
        rate_usage:
          type: object
          additionalProperties:
            type: integer
            format: int64
        environment:
          type: string
        api_scope:
          type: string
    ActionEnvelope:
      type: object
      required:
        - action_id
        - agent_id
        - capability_id
        - action_type
        - action_payload
        - constraint_evidence
        - timestamp
        - agent_signature
      properties:
        action_id:
          type: string
        agent_id:
          type: string
        capability_id:
          type: string
        action_type:
          type: string
        action_payload:
          type: object
        constraint_evidence:
          $ref: '#/components/schemas/ConstraintEvidence'
        timestamp:
          type: string
          format: date-time
        agent_signature:
          type: string
    IssueCapabilityRequest:
      type: object
      required:
        - agent_public_key
        - allowed_actions
        - constraints
        - expires_at
      properties:
        agent_public_key:
          type: string
        allowed_actions:
          type: array
          items:
            type: string
        constraints:
          $ref: '#/components/schemas/CapabilityConstraints'
        expires_at:
          type: string
          format: date-time
        nonce:
          type: string
    IssueCapabilityResponse:
      type: object
      required:
        - capability
        - issuer
      properties:
        capability:
          $ref: '#/components/schemas/Capability'
        issuer:
          type: object
          required:
            - issuer_id
            - public_key
          properties:
            issuer_id:
              type: string
            public_key:
              type: string
    VerifyActionRequest:
      type: object
      required:
        - capability
        - action
        - issuer_public_key
        - agent_public_key
        - revoked_capability_ids
      properties:
        capability:
          $ref: '#/components/schemas/Capability'
        action:
          $ref: '#/components/schemas/ActionEnvelope'
        issuer_public_key:
          type: string
        agent_public_key:
          type: string
        revoked_capability_ids:
          type: array
          items:
            type: string
    VerificationResult:
      type: object
      required:
        - decision
        - reasons
      properties:
        decision:
          type: string
          enum: [AUTHORIZED, REJECTED]
        reasons:
          type: array
          items:
            type: string